requirements.txt
streamlit
langchain
langchain-groq
langchain-community
pypdf
chromadb
faiss-cpu
sentence-transformers
python-dotenv
langchain-openai